Team-A BillBaord Hit Prediction(LR&GDA)/billboard-hitpredictor-streamlit-heroku/requirements.txt
scikit-learn
numpy
pandas
matplotlib
streamlit
Pillow
seaborn